requirements.txt
python-binance==1.0.19
scikit-learn==1.5.1
numpy==2.0.1
pandas==2.2.2
tqdm==4.66.5